requirements.txt
asgiref==3.5.2
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
Django==3.2.16
djangorestframework==3.14.0
pytz==2022.6
sqlparse==0.4.3
typing_extensions==4.4.0
wincertstore==0.2
